@@ -1,18 +1,17 @@
 [build-system]
-requires = ["setuptools>=45", "setuptools_scm[toml]>=6.2"]
-build-backend = "setuptools.build_meta"
+requires = ["setuptools>=60", "setuptools-scm>=8.0"]
 
 [tool.setuptools_scm]
-write_to = "_version.py"
+version_file="_version.py"
 
 [tool.black]
 line-length = 95
-target-version = ["py310"]
+target-version = ["py311"]
 extend-exclude = '''^(.*\/)*\b(migrations)\b($|\/.*$)'''
 
 [tool.isort]
 profile = "black"
-py_version = "310"
+py_version = "311"
 skip = [".tox", ".eggs", "migrations"]
 
 [tool.coverage.run]
